add lock file for caching + clear actions that are unsuccesful at the end
